refactor(settings): success_rate_penalty_weight 类型 float → int
整数类型更适合此配置项的实际用途，用户只会设置 50、100、200 等整数值。
包含数据库迁移，自动更新已有记录的 value_type。

+++ internal/app/selector_balancer.go
@@ -104,7 +104,7 @@ func (s *Server) calculateEffectivePriority(
 	}
 
 	// 惩罚 = 失败率 × 权重 × 置信度
-	penalty := failureRate * cfg.SuccessRatePenaltyWeight * confidence
+	penalty := failureRate * float64(cfg.SuccessRatePenaltyWeight) * confidence
 
 	return basePriority - penalty
 }
